// src/api/generated/model/reason.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * LeadRadar API
 * OpenAPI spec version: 0.1.0
 */
import type { ReasonPolarity } from './reasonPolarity';

export interface Reason {
  text: string;
  polarity: ReasonPolarity;
  signal_id: string | null;
  source_name: string | null;
  url: string | null;
  date: string | null;
}